pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "Playing around with MCP agents"
authors = [
    { name = "fred", email = "fred@fred.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]",
    "mcp-server",
    "requests>=2.32.3",
    "python-dotenv>=1.1.0",
    "notion-client>=2.3.0",
    "nest-asyncio>=1.6.0",
    "openai>=1.84.0",
    "pre-commit>=4.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
    "pre-commit",
    "isort"
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
